Patch SVF ExtAPI functionality to use hardcoded path in Nix store
SVF includes an "ExtAPI.json" file that contains a database of known functions
and how to handle them. It looks them up by 1) looking in the optionally
provided argument; 2) searching for a hardcoded path; 3) looking in an
environment variable; or 4) calling `npm` to determine where SVF was installed.

This is undesirable for us, so SVF is instead patched so that the ExtAPI.json is
installed to a hardcoded path in the Nix store and looked up from there.